Added a validation for the admin page.

+++ server/controllers/admin.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const formidable = require('formidable');
 const db = require('../models/db');
+const validation = require('../libs/validation');
 
 module.exports.get = function(req, res) {
   if (req.session.isAdmin) {
@@ -19,9 +20,16 @@ module.exports.get = function(req, res) {
 
 module.exports.postSkills = function(req, res) {
   const { age, concerts, cities, years } = req.body;
-  db.set('skills', { age, concerts, cities, years }).write();
-  req.flash('info', 'Навыки обновлены');
-  res.redirect('/admin');
+  const errors = validation.skills({ age, concerts, cities, years });
+
+  if (errors.length) {
+    req.flash('info', errors.join('. '));
+    res.redirect('/admin');
+  } else {
+    db.set('skills', { age, concerts, cities, years }).write();
+    req.flash('info', 'Навыки обновлены');
+    res.redirect('/admin');
+  }
 };
 
 module.exports.postUpload = function(req, res) {
@@ -36,23 +44,35 @@ module.exports.postUpload = function(req, res) {
 
   form.parse(req, function(err, fields, files) {
     if (err) {
-      fs.unlinkSync(files.photo.path);
+      if (fs.existsSync(files.photo.path)) {
+        fs.unlinkSync(files.photo.path);
+      }
       req.flash('form-info', 'Возникла ошибка при обработке');
       res.redirect('/admin');
     }
 
-    const fileName = path.join(upload, files.photo.name);
-    fs.renameSync(files.photo.path, fileName);
-    // removes the server\\public a part of the path
-    const dir = fileName.replace('server\\public', '');
+    const { name: title, price } = fields;
+    const { name, size } = files.photo;
+    const errors = validation.uploadFile({ name, size, title, price });
+    if (errors.length) {
+      if (fs.existsSync(files.photo.path)) {
+        fs.unlinkSync(files.photo.path);
+      }
+      req.flash('form-info', errors.join('. '));
+      res.redirect('/admin');
+    } else {
+      const fileName = path.join(upload, files.photo.name);
+      fs.renameSync(files.photo.path, fileName);
+      // removes the server\\public a part of the path
+      const dir = fileName.replace('server\\public', '');
 
-    const { name, price } = fields;
-    db.defaults({ products: [] })
-      .get('products')
-      .push({ src: dir, name, price })
-      .write();
+      db.defaults({ products: [] })
+        .get('products')
+        .push({ src: dir, name: title, price })
+        .write();
 
-    req.flash('form-info', 'Форма обработана');
-    res.redirect('/admin');
+      req.flash('form-info', 'Форма обработана');
+      res.redirect('/admin');
+    }
   });
 };
